add master slave datasource and refactor the module name

+++ sharding-jdbc-master-slave-jpa-example/src/main/java/sharding/jdbc/example/model/Order.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
 import java.sql.Timestamp;
 
 import javax.persistence.Column;
@@ -20,11 +19,11 @@ import javax.persistence.Id;
 public class Order {
 
     private Integer tenantId;
-    private BigInteger userId;
+    private long userId;
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "order_id")
-    private Long orderId;
+    private long orderId;
     private String orderName;
     private Timestamp createTime;
 }
